test(precommit): isolate check and push ordering
